Change Google Blogger credential storage to use explicit FileDataStore
Most blog APIs take a simple username/password pair and our
IBlogCredentials are based around this concept. For OAuth, there is more
to store than just a simple username/password. The Google APIs already
provide an option to automatically handle the data storage for the
OAuth, so I'm using that instead.