List of recent updated posts was pointing to the old folder
The problem was that the list of files recently worked on was cached and
not based on the  new folder that was updated in option.

I was working on a more OO method but got tangled in a bunch of
reference issues.

I did find an existing  method for updating the list recent files it
does rely on the relationship between Forms, Controls and Parent
objects.