Show drafts and scheduled posts and pages for Google Blogger
We weren't including the correct API parameters to request drafts and
scheduled posts from Google Blogger. When getting the list of posts and
pages, we need to do 2-3 requests with different parameters and combine
the results. For getting individual posts and pages, we had to specify
view permissions in the request.